--- src/main/java/service/RunRace.java
@@ -1,20 +1,0 @@
-package service;
-
-import domain.Cars;
-import domain.RaceRecord;
-
-public class RunRace {
-    private final int tryCount;
-
-    public RunRace(int tryCount) {
-        this.tryCount = tryCount;
-    }
-
-    public RaceRecord run(RaceRecord raceRecord, Cars cars) {
-        for (int count = 0; count < tryCount; count++) {
-            raceRecord.addRecord(cars.move());
-        }
-        return raceRecord;
-    }
-
-}
