+++ src/test/java/domain/RunRaceTest.java
@@ -1,6 +1,5 @@
 package domain;
 
-import service.RunRace;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -12,11 +11,8 @@ class RunRaceTest {
     @Order(2)
     @ValueSource(ints = {3, 4, 5})
     void 주어진_횟수만큼_레이스가_진행된다(int tryCount) {
-        Cars cars = Cars.builder()
-                .carsName("car,truck")
-                .raceCondition(new RaceCondition())
-                .build();
-
-        assertThat(new RunRace(tryCount).run(new RaceRecord(), cars).getSize()).isEqualTo(tryCount);
+        assertThat(new RunRace().run("car,truck", tryCount)
+                .getSize())
+                .isEqualTo(tryCount);
     }
 }
